Add initial support for constant-sized arrays
Arrays can be declared using
```
    let <var_name>[<length>] = { <arith_expr>, ... };
```
syntax. # of elements in curly braces must not be greater than
length (however, this is not currently checked) - these will be values
of arrays' elements. If # of elements is smaller than length,
missing indices will get value as if element's list was cyclical.

Various other expressions in brackets are considered to be added ex.
range selection - <var_name>[beg:end] and/or <var_name>[1,5,10].